Move the ttkbootstrap theme support for matplotlib to its own module
